backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic-settings>=2.2
openai>=1.0